requirements.txt
Django==3.2
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.2.0
pytest==7.1.2
